backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
reportlab==4.2.5
sqlparse==0.5.3
typing_extensions==4.12.2
